requirements.txt
streamlit
pandas
numpy
requests
geopy
pydeck
scikit-learn
openai
python-dotenv
